* Conversion refactoring part2
