+++ com.dotsui.hybrid/DotsUI.Hybrid/ConvertToEntity/Assets.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ae002e213eb1f0d4fa1d9778d8f73840
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
