+++ com.dotsui.hybrid/DotsUI.Hybrid/ConvertToEntity/Assets/DeclareReferencedSpriteAssetSystem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a8bd8cd149ccd4544b2b5ee5c5fcc90c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
