+++ com.dotsui.hybrid/DotsUI.Hybrid/ConvertToEntity/Assets/DeclareReferencedTMPFontAssetSystem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 21d9ef9c1858e1f40acad78b3657384c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
