+++ com.dotsui.hybrid/DotsUI.Hybrid/ConvertToEntity/Assets/LegacyTextFontAssetProxy.cs
@@ -1,0 +1,8 @@
+﻿using System;
+using Unity.Entities;
+
+namespace DotsUI.Core
+{
+	[Serializable]
+	class LegacyTextFontAssetProxy : SharedComponentDataProxy<LegacyTextFontAsset> { }
+}
