+++ com.dotsui.hybrid/DotsUI.Hybrid/ConvertToEntity/Assets/LegacyTextFontAssetProxy.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7e9fd0d9695ed5c44b32b15424cf0d5e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
