+++ com.dotsui.hybrid/DotsUI.Hybrid/ConvertToEntity/Assets/SpriteAssetConversionSystem.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: fe12e80f498b6554482b8277cbbe7e91
+guid: 862180fd454c0144bade9d9993541750
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
