+++ com.dotsui.hybrid/DotsUI.Hybrid/ConvertToEntity/Assets/SpriteAssetProxy.cs
@@ -1,0 +1,8 @@
+﻿using System;
+using Unity.Entities;
+
+namespace DotsUI.Core
+{
+	[Serializable]
+	class SpriteAssetProxy : SharedComponentDataProxy<SpriteAsset> { }
+}
