+++ com.dotsui.hybrid/DotsUI.Hybrid/ConvertToEntity/Assets/SpriteAssetProxy.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 89cee1cc020e46e429df335a9ce7226f
+guid: 13d4b20d7e65ec043b6cebdaa060f101
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
