+++ com.dotsui.hybrid/DotsUI.Hybrid/ConvertToEntity/Assets/TMPFontAssetConversionSystem.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: de9b6798f91e4dc45b5d5e974cf5b786
+guid: f00e23a9aa566fd47b1516216073b492
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
