+++ com.dotsui.hybrid/DotsUI.Hybrid/ConvertToEntity/CanvasConversionSystem.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e249074e4fc412340a613dd9eda5e8e5
+guid: 9582357aceef8a744aaff9d112b7fa58
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
