+++ com.dotsui.hybrid/DotsUI.Hybrid/ConvertToEntity/Controls/ScrollBarConversionSystem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 91c150e7af385ee48bb372194074d5e5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
