--- com.dotsui.hybrid/DotsUI.Hybrid/ConvertToEntity/LegacyTextFontAssetProxy.cs
@@ -1,8 +1,0 @@
-﻿using System;
-using Unity.Entities;
-
-namespace DotsUI.Core
-{
-	[Serializable]
-	class LegacyTextFontAssetProxy : SharedComponentDataProxy<LegacyTextFontAsset> { }
-}
