--- com.dotsui.hybrid/DotsUI.Hybrid/ConvertToEntity/SpriteAssetProxy.cs
@@ -1,8 +1,0 @@
-﻿using System;
-using Unity.Entities;
-
-namespace DotsUI.Core
-{
-	[Serializable]
-	class SpriteAssetProxy : SharedComponentDataProxy<SpriteAsset> { }
-}
